requirements.txt
twitchio==2.10.0
langchain-core~=0.3.65
langgraph~=0.4.8
langmem~=0.0.27
redis~=6.2.0
python-dotenv~=1.1.0
numpy~=2.3.0
matplotlib~=3.10.3
scipy~=1.15.3